refactor: refactor barrels in authentication
