+++ src/features/authentication/index.ts
@@ -1,13 +1,5 @@
-export { useResendOneTimePassword } from './confirm-user/model/use-resend-one-time-password'
-export { useSendOneTimePassword } from './confirm-user/model/use-send-one-time-password'
 export { default as authDe } from './i18n/authDe'
 export { default as authEn } from './i18n/authEn'
 export { default as translationKeys } from './i18n/keys'
 export { useAuthNavigation } from './model/use-auth-navigation'
 export { useAuthSession } from './model/use-auth-session'
-export { default as translationKeysPasswordForgotten } from './password-forgotten/i18n/keys'
-export { default as OtpStep } from './password-forgotten/ui/OtpStep.vue'
-export { default as PasswordStepper } from './password-forgotten/ui/PasswordStepper.vue'
-export { default as RegisterForm } from './register-user/ui/RegisterForm.vue'
-export { default as LoginForm } from './signin-user/ui/LoginForm.vue'
-export { default as LocalWork } from './work-local/ui/LocalWork.vue'
