+++ src/features/authentication/password-forgotten/index.ts
@@ -1,0 +1,2 @@
+export { default as translationKeys } from './i18n/keys'
+export { default as PasswordStepper } from './ui/PasswordStepper.vue'
