+++ src/features/authentication/register-user/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterForm } from './ui/RegisterForm.vue'
