+++ src/features/authentication/signin-user/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginForm } from './ui/LoginForm.vue'
