+++ src/features/authentication/work-local/index.ts
@@ -1,0 +1,1 @@
+export { default as LocalWork } from './ui/LocalWork.vue'
