+++ src/pages/dashboard/ui/Welcome.vue
@@ -1,5 +1,3 @@
-<script setup lang="ts"></script>
-
 <template>
   <v-container class="fill-height d-flex align-center justify-center">
     <v-card class="pa-8" max-width="480" width="100%">
